requirements.txt
future==0.17.1
grpcio==1.17.1
grpcio-tools==1.17.1
numpy==1.15.4
pandas==0.24.1
scikit-learn==0.20.2
scipy==1.2.1
